Reads file and does calculations
Reads text from file and does the calculations. Only thing left to do is decide how we want the user to enter the file they would like to use. Either through command line prompts or otherwise. Once we decide we just need to implement that and it's good to go